Fix generator bug: completed PoC for generate,exercise, evaluate
The problem was that the random offset implies there is no
warranty that each window will have a record with the window
start as timestamp. That is ok, and supports empty windows.
Fixed by adding a explicit parameter for the first window start
time to both TestCaseGenerator.batchesToStream and
TimedValue.tumblingWindows, so window generation and evaluation
use the same windows